+++ packages/hermes/src/email/index.ts
@@ -1,0 +1,1 @@
+export { default as emailNotifier } from './sendEmail'