+++ packages/is/src/isPasswordLength.js
@@ -1,0 +1,11 @@
+/**
+ * @param {string} value The value to be evaluated
+ * @returns {boolean} Returns true if the value's length is greater than 8
+ *
+ * @description
+ * Checks whether the given value has is at least 8 characters long
+ */
+
+const isPasswordLength = (value) => value && value.length >= 8
+
+export default isPasswordLength