+++ packages/function/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/beatgig/midi/compare/@beatgig/function@1.0.2...@beatgig/function@1.0.3) (2020-02-04)
+
+**Note:** Version bump only for package @beatgig/function
+
+
+
+
+
 ## [1.0.2](https://github.com/beatgig/midi/compare/@beatgig/function@1.0.1...@beatgig/function@1.0.2) (2019-09-01)
 
 **Note:** Version bump only for package @beatgig/function